fix: 로그인 상태에 따라서 헤더 수정
- 로고 누르면 강의실로 이동
- 강의실일 때는 '대시보드로 가기' / 대시보드일 때는 '강의실로 가기'
